réglage de l'affichage de la vue postsbytopic en tant qu'utilisateur lambda

+++ view/forum/PostsByTopics.php
@@ -131,13 +131,14 @@ $isBan = (isset($_SESSION["user"]) && ($_SESSION["user"]->getBannedUntil() == NU
         
         <!-- Affiche le bouton supprimer uniquement si $nombrePosts et supérieur a 1 -->
         <?php if ($nombrePosts > 1) { ?>
-
-            <?php 
-            // On vérifie que l'utilisateur en session est soit un admin ou l'auteur du sujet pour permettre la suppression d'un post
-            if($isAdmin || $isSession && $_SESSION["user"]->getId() == $post->getUser()->getId()) { ?>
-                <a class="btn-delete" href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>">Supprimer</a>
+            <?php
+                // On vérifie que l'auteur du post existe
+                if(($post->getUser())) {
+                    // On vérifie que l'utilisateur en session est soit un admin ou l'auteur du post pour permettre la suppression d'un post
+                    if($isAdmin || $isSession && $_SESSION["user"]->getId() == $post->getUser()->getId()) { ?>
+                        <a class="btn-delete" href="index.php?ctrl=forum&action=deletePost&id=<?= $post->getId() ?>">Supprimer</a>
+                    <?php } ?>
             <?php } ?>
-            
         <?php } ?>
     </div>
     <?php } ?>
